Ref(textual): Move agent to interactive_textual
